requirements.txt
django
django-phonenumber-field
django-phonenumbers